refactor: #52-folder-tree 트리 조회 Row DTO를 repository로 일원화
