--- src/main/java/com/keepgoing/keepgoing/folder/service/dto/FolderTreeRow.java
@@ -1,8 +1,0 @@
-package com.keepgoing.keepgoing.folder.service.dto;
-
-public record FolderTreeRow(
-        Long folderId,
-        Long parentId,
-        String name
-) {
-}
